feat: allow block element as item’s list item view
Currently it’s useful for showing RTL data, since we can’t put `dir` on
list item’s container then list item container must be a block in order
for RTL content to start really from the right.

Not completely sure about this solution, we could perhaps find a way for
list item view to pass `dir` to container instead, but blocky list item
views could also be handy in case of images so perhaps it’s the way to
go.